feat: 메인 페이지 기능 통합 및 스크롤 애니메이션 통합 (미완성)
